chore(test): move schedule wrapper under test file
