+++ pkg/cmd/schedule.go
@@ -358,42 +358,9 @@ func checkScheduleschedule(crontab, apiKey, url, fronius_ip string, pw_consumpti
 	return nil
 }
 
-// schedule performs a single execution of the scheduling workflow:
-//   - retrieves the forecast
-//   - reads storage state
-//   - computes the Fronius decision
-//   - publishes the state over MQTT (if configured)
-//
-// It is intentionally retained as a compatibility wrapper for tests.
-// New runtime paths should call Runner.Tick directly.
-func schedule(apiKey, url, fronius_ip string, pw_consumption, max_charge, pw_batt_reserve float64,
-	start_hr, end_hr, batt_reserve_start_hr, batt_reserve_end_hr string, pw_lwt, pw_upt float64,
-	cache_forecast bool, cache_file_prefix string, cache_time int32, mqttClient mqtt.Client, mqttCfg mqtt.Config) {
-	runnerCfg := RunnerConfig{
-		APIKey:             apiKey,
-		URL:                url,
-		FroniusIP:          fronius_ip,
-		PWConsumption:      pw_consumption,
-		MaxCharge:          max_charge,
-		PWBattReserve:      pw_batt_reserve,
-		StartHR:            start_hr,
-		EndHR:              end_hr,
-		BattReserveStartHR: batt_reserve_start_hr,
-		BattReserveEndHR:   batt_reserve_end_hr,
-		PWLWT:              pw_lwt,
-		PWUPT:              pw_upt,
-		CacheForecast:      cache_forecast,
-		CacheFilePrefix:    cache_file_prefix,
-		CacheTime:          cache_time,
-		MQTT:               mqttCfg,
-		Now:                time.Now,
-	}
-
-	runner := NewRunner(runnerCfg, mqttClient)
-	if err := runner.Tick(context.Background(), time.Now()); err != nil {
-		u.Log.Error(err)
-	}
-}
+// NOTE: the single-shot `schedule` compatibility wrapper was moved to
+// package tests (pkg/cmd/schedule_test.go). Tests should call the
+// wrapper or directly call `NewRunner(...).Tick(...)`.
 
 // publishStateSnapshot publishes the provided `payload` using the MQTT client
 // and stores a copy in `latestState` (if provided). The copy is used to
